AG-13658 - Refactor framework text animation to only include animated content
