+++ documentation/ag-grid-docs/src/pages-styles/homepage.module.scss
@@ -33,6 +33,16 @@ body {
     }
 }
 
+.heroInner {
+    display: flex;
+    flex-direction: column;
+    gap: 0 $spacing-size-16;
+
+    @media screen and (min-width: $breakpoint-hero-large) {
+        flex-direction: row;
+    }
+}
+
 .heroHeadings {
     flex: 1;
     display: flex;
@@ -84,14 +94,8 @@ body {
     }
 }
 
-.heroInner {
-    display: flex;
-    flex-direction: column;
-    gap: 0 $spacing-size-16;
-
-    @media screen and (min-width: $breakpoint-hero-large) {
-        flex-direction: row;
-    }
+.heroTopLine {
+    display: inline-flex;
 }
 
 .heroSectionDemoContainer {
