+++ external/ag-website-shared/src/components/framework-text-animation/FrameworkTextAnimation.module.scss
@@ -1,9 +1,5 @@
 @use 'design-system' as *;
 
-.topLine {
-    display: inline-flex;
-}
-
 .animatedWordsOuter {
     --word-height: 1.3em;
     --word-index: 0;
